+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/cipher/CipherDecryptCommand.java
@@ -6,10 +6,13 @@ import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import org.thermoweb.intellij.plugin.encrypt.Algorithms;
 import org.thermoweb.intellij.plugin.encrypt.CipherInformationsDialog;
 import org.thermoweb.intellij.plugin.encrypt.JasyptPluginSettings;
 import org.thermoweb.intellij.plugin.encrypt.Notifier;
 import org.thermoweb.intellij.plugin.encrypt.exceptions.JasyptPluginException;
+import org.thermoweb.intellij.plugin.encrypt.vault.CipherConfiguration;
+import org.thermoweb.intellij.plugin.encrypt.vault.SecretVault;
 
 import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.psi.PsiLanguageInjectionHost;
@@ -28,19 +31,60 @@ public class CipherDecryptCommand {
         return new CipherDecryptCommand(property);
     }
 
+    public boolean check() {
+        boolean isTextEncapsulated = Optional.ofNullable(property.getText())
+                .map(String::trim)
+                .map(s -> s.startsWith("ENC("))
+                .orElse(false);
+        Optional<CipherConfiguration> storedConfiguration = SecretVault.getSecrets(property.getContainingFile().getVirtualFile().getPath());
+        if (storedConfiguration.isEmpty()) {
+            return false;
+        }
+
+        CipherConfiguration configuration = storedConfiguration.get();
+        String textToUncrypt = getTextToUncrypt(property.getText(), isTextEncapsulated);
+        try {
+            CipherUtils.decrypt(textToUncrypt, configuration.password(), configuration.algorithm().getCode());
+            return true;
+        } catch (JasyptPluginException e) {
+            return false;
+        }
+    }
+
     public void execute() {
         boolean isTextEncapsulated = Optional.ofNullable(property.getText())
                 .map(String::trim)
                 .map(s -> s.startsWith("ENC("))
                 .orElse(false);
+        Optional<CipherConfiguration> storedConfiguration = SecretVault.getSecrets(property.getContainingFile().getVirtualFile().getPath());
+        storedConfiguration.ifPresentOrElse(configuration -> {
+
+                    String textToUncrypt = getTextToUncrypt(property.getText(), isTextEncapsulated);
+                    try {
+                        String clearText = CipherUtils.decrypt(textToUncrypt, configuration.password(), configuration.algorithm().getCode());
+                        setClearText(clearText);
+                    } catch (JasyptPluginException e) {
+                        askAndDecrypt(isTextEncapsulated);
+                    }
+                },
+                () -> askAndDecrypt(isTextEncapsulated));
+    }
+
+    private void askAndDecrypt(boolean isTextEncapsulated) {
         CipherInformationsDialog dialog = new CipherInformationsDialog(JasyptPluginSettings.getInstance(property.getProject()), isTextEncapsulated);
         if (!dialog.showAndGet()) {
             return;
         }
         Map<String, String> values = dialog.getValues();
+        String textToUncrypt = getTextToUncrypt(property.getText(), values.get(ENCAPSULATE_FIELD_NAME));
+        String password = values.get(PASSWORD_FIELD_NAME);
+        String algorithm = values.get(ALGORITHM_FIELD_NAME);
         try {
-            String textToUncrypt = getTextToUncrypt(property.getText(), values.get(ENCAPSULATE_FIELD_NAME));
-            String clearText = CipherUtils.decrypt(textToUncrypt, values.get(PASSWORD_FIELD_NAME), values.get(ALGORITHM_FIELD_NAME));
+            String clearText = CipherUtils.decrypt(textToUncrypt, password, algorithm);
+            if ("true".equals(values.get(REMEMBER_PASSWORD))) {
+                SecretVault.storeSecret(property.getContainingFile().getVirtualFile().getPath(), new CipherConfiguration(Algorithms.fromCode(algorithm), password));
+            }
+            JasyptPluginSettings.updateSettings(property.getProject(), dialog.getValues());
             setClearText(clearText);
         } catch (JasyptPluginException e) {
             Notifier.notifyError(property.getProject(), "Failed to decrypt string, please verify provided password or algorithm.");
@@ -48,8 +92,12 @@ public class CipherDecryptCommand {
     }
 
     private String getTextToUncrypt(String text, String isEncapsulated) {
+        return getTextToUncrypt(text, "true".equals(isEncapsulated));
+    }
+
+    private String getTextToUncrypt(String text, boolean isEncapsulated) {
         Matcher matches = pattern.matcher(text);
-        return matches.find() && "true".equals(isEncapsulated) ?
+        return matches.find() && isEncapsulated ?
                 matches.group(1) :
                 text;
     }